Refactor: 차트 옵션 객체 분리
